+++ appData.js
@@ -1,4 +1,4 @@
-const logs = {
+export const logs = {
 	start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
 	end: [
 			'Результат удара [playerWins]: [playerLose] - труп',
@@ -37,3 +37,10 @@ const logs = {
 	],
 	draw: 'Ничья - это тоже победа!'
 };
+
+export const HIT = {
+	head: 30,
+	body: 25,
+	foot: 20,
+};
+export const ATTACK = ['head', 'body', 'foot'];
